+++ crates/ironcalc-base/src/expressions/parser/static_analysis.rs
@@ -929,6 +929,9 @@ fn get_function_args_signature(kind: &Function, arg_count: usize) -> Vec<Signatu
         Function::Choosecols | Function::Chooserows | Function::Sortby => {
             vec![Signature::Vector; arg_count]
         }
+        Function::Mmult | Function::Minverse | Function::Munit => {
+            vec![Signature::Vector; arg_count]
+        }
         Function::Tocol | Function::Torow | Function::Expand => {
             let mut result = vec![Signature::Vector; arg_count];
             for item in result.iter_mut().skip(1) {
@@ -1539,6 +1542,9 @@ fn static_analysis_on_function(kind: &Function, args: &[Node]) -> StaticResult {
         Function::Torow => not_implemented(args),
         Function::Expand => not_implemented(args),
         Function::Sortby => not_implemented(args),
+        Function::Mmult => not_implemented(args),
+        Function::Minverse => not_implemented(args),
+        Function::Munit => not_implemented(args),
         Function::Aggregate => not_implemented(args),
         Function::Rand => not_implemented(args),
         Function::Randbetween => scalar_arguments(args),