+++ crates/ironcalc-base/src/functions/mod.rs
@@ -546,6 +546,9 @@ pub enum Function {
     Torow,
     Expand,
     Sortby,
+    Mmult,
+    Minverse,
+    Munit,
 }
 
 macro_rules! impl_function_lookup {
@@ -1066,6 +1069,9 @@ impl_function_lookup! {
     torow           => Torow,
     expand          => Expand,
     sortby          => Sortby,
+    mmult           => Mmult,
+    minverse        => Minverse,
+    munit           => Munit,
 }
 
 pub fn english_function_names() -> Vec<String> {
@@ -1551,9 +1557,12 @@ impl Function {
             Function::Torow => functions.torow.clone(),
             Function::Expand => functions.expand.clone(),
             Function::Sortby => functions.sortby.clone(),
+            Function::Mmult => functions.mmult.clone(),
+            Function::Minverse => functions.minverse.clone(),
+            Function::Munit => functions.munit.clone(),
         }
     }
-    pub fn into_iter() -> IntoIter<Function, 469> {
+    pub fn into_iter() -> IntoIter<Function, 472> {
         [
             Function::And,
             Function::False,
@@ -2024,6 +2033,9 @@ impl Function {
             Function::Torow,
             Function::Expand,
             Function::Sortby,
+            Function::Mmult,
+            Function::Minverse,
+            Function::Munit,
         ]
         .into_iter()
     }
@@ -2084,6 +2096,7 @@ impl Function {
             Function::Torow => "_xlfn.TOROW".to_string(),
             Function::Expand => "_xlfn.EXPAND".to_string(),
             Function::Sortby => "_xlfn.SORTBY".to_string(),
+            Function::Munit => "_xlfn.MUNIT".to_string(),
             Function::Isformula => "_xlfn.ISFORMULA".to_string(),
             Function::Sheet => "_xlfn.SHEET".to_string(),
             Function::Formulatext => "_xlfn.FORMULATEXT".to_string(),
@@ -2697,6 +2710,9 @@ impl<'a> Model<'a> {
             Function::Torow => self.fn_torow(args, cell),
             Function::Expand => self.fn_expand(args, cell),
             Function::Sortby => self.fn_sortby(args, cell),
+            Function::Mmult => self.fn_mmult(args, cell),
+            Function::Minverse => self.fn_minverse(args, cell),
+            Function::Munit => self.fn_munit(args, cell),
             Function::Gauss => self.fn_gauss(args, cell),
             Function::Harmean => self.fn_harmean(args, cell),
             Function::Kurt => self.fn_kurt(args, cell),